Extract document loading skeleton to standalone, reusable component
